Be more careful about what gets opened in new windows
Only open the resource in a new window if the navigation type was
NavigationType.LINK_CLICKED.

Also, do not assume that the WebKitPolicyDecision has a get_request()
method. That's true for both WebKitNavigationPolicyDecision and
WebKitResponsePolicyDecision, the implementations of
WebKitPolicyDecision that exist today, but it might not be true of
future WebKitPolicyDecisions.