views: correctly delegate the view layout
The GenericFeedsView abstract class currently adapts
the layout according to the show_feedlist argument. It
turns out that this is a poor design choice, and a better
and more sustainable path is to delegate the layout
setup to the subclass.

Fix this by moving the layout code to a separate function
and override this function in FeedsView.